modify overtime

+++ common.h
@@ -49,7 +49,7 @@ inline long GetCurrentTimeMs(void)
 {
     struct timeval tv;
     gettimeofday(&tv,NULL);
-    return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    return (int64_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
 }
 
 inline string GetCurrentTimeTiff(void)
